Separeted in diferents files laboratory-request dto and create-microbiology-request dto | Aplied prettier
