+++ src/laboratory-request/dto/update-laboratory-request.dto.ts
@@ -1,4 +1,6 @@
 import { PartialType } from '@nestjs/swagger';
 import { CreateLaboratoryRequestDto } from './create-laboratory-request.dto';
 
-export class UpdateLaboratoryRequestDto extends PartialType(CreateLaboratoryRequestDto) {}
+export class UpdateLaboratoryRequestDto extends PartialType(
+  CreateLaboratoryRequestDto,
+) {}
